refactor: memory_converter를 memory_modules로 분리하고 동작/포맷은 그대로 유지
